refactor(web): make settings an inline route instead of a modal
SettingsDialog fused four jobs into 1,571 lines: the modal shell, the tab nav,
ten page bodies and their derived state. Because the shell was welded to the
content, the pages could only ever render as an overlay.

Split it. useSettingsNav owns the tab union, the groups, the active tab and the
guarded lazy loads; SettingsNav renders the nav inside the real sidebar, and
SettingsPane hosts one component per page in the content area. The sidebar keeps
its header and New Session button in both modes and swaps only its body.
