+++ apps/pythinker-web/src/components/Sidebar.vue
@@ -6,8 +6,10 @@
 import { nextTick, onBeforeUnmount, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import type { Session, WorkspaceGroup, WorkspaceView } from '../types';
+import type { SettingsTab } from '../composables/useSettingsNav';
 import SessionRow from './SessionRow.vue';
 import PythinkerLogo from './PythinkerLogo.vue';
+import SettingsNav from './settings/SettingsNav.vue';
 
 const { t } = useI18n();
 
@@ -24,6 +26,8 @@ const props = withDefaults(
     unreadBySession?: Record<string, boolean>;
     /** Width (px) of the session column, driven by the App resize handle. */
     colWidth?: number;
+    mode?: 'sessions' | 'settings';
+    activeSettingsTab?: SettingsTab;
   }>(),
   {
     activeWorkspace: null,
@@ -32,6 +36,8 @@ const props = withDefaults(
     pendingBySession: () => ({}),
     unreadBySession: () => ({}),
     colWidth: 220,
+    mode: 'sessions',
+    activeSettingsTab: 'general',
   },
 );
 
@@ -48,6 +54,8 @@ const emit = defineEmits<{
   renameWorkspace: [id: string, name: string];
   deleteWorkspace: [id: string];
   openSettings: [];
+  closeSettings: [];
+  selectSettingsTab: [tab: SettingsTab];
   collapse: [];
 }>();
 
@@ -485,6 +493,9 @@ onBeforeUnmount(() => {
         </button>
       </div>
 
+      <!-- The header and New Session button above stay mounted in both modes; only
+           the body below swaps between the session list and the settings nav. -->
+      <template v-if="mode === 'sessions'">
       <!-- Workspace section actions -->
       <div class="ws-head">
         <input
@@ -666,6 +677,21 @@ onBeforeUnmount(() => {
           <span>{{ t('settings.title') }}</span>
         </button>
       </div>
+      </template>
+
+      <template v-else>
+        <div class="settings-nav-body">
+          <SettingsNav :active-tab="activeSettingsTab" @select="emit('selectSettingsTab', $event)" />
+        </div>
+        <div class="side-foot">
+          <button type="button" class="settings-row" @click.stop="emit('closeSettings')">
+            <svg viewBox="0 0 16 16" width="16" height="16" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
+              <path d="m10 3-5 5 5 5" />
+            </svg>
+            <span>{{ t('settings.backToSessions') }}</span>
+          </button>
+        </div>
+      </template>
     </div>
 
     <div
@@ -1150,6 +1176,11 @@ onBeforeUnmount(() => {
 }
 
 /* Pinned settings action */
+.settings-nav-body {
+  display: flex;
+  flex: 1;
+  min-height: 0;
+}
 .side-foot {
   flex: none;
   border-top: none;
