+++ apps/pythinker-web/src/components/settings/settings.css
@@ -1,0 +1,260 @@
+.panel { display: block; }
+.sec { padding: 12px 0; border-bottom: 1px solid var(--line); }
+.sec:last-child { border-bottom: none; }
+.sec-title {
+  margin: 0 0 10px;
+  color: var(--muted);
+  font-family: var(--mono);
+  font-size: calc(var(--ui-font-size) - 3px);
+  font-weight: 700;
+  letter-spacing: 0.06em;
+  text-transform: uppercase;
+}
+.sec-note {
+  margin: -4px 0 12px;
+  color: var(--muted);
+  font-size: calc(var(--ui-font-size) - 2px);
+}
+.sec-empty {
+  margin: 0;
+  color: var(--faint);
+  font-size: calc(var(--ui-font-size) - 1px);
+}
+.row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  min-height: 34px;
+  padding: 3px 0;
+}
+.rlabel {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  color: var(--ink);
+  font-family: var(--sans);
+  font-size: calc(var(--ui-font-size) - 0.5px);
+}
+.rvalue {
+  max-width: 60%;
+  overflow: hidden;
+  color: var(--muted);
+  font-family: var(--sans);
+  font-size: calc(var(--ui-font-size) - 1.5px);
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+.rvalue.mono { font-family: var(--mono); font-size: var(--ui-font-size-xs); }
+.hint { color: var(--faint); font-family: var(--sans); font-size: calc(var(--ui-font-size) - 3px); }
+.act {
+  padding: 6px 12px;
+  border: 1px solid var(--line);
+  border-radius: 7px;
+  background: var(--bg);
+  color: var(--ink);
+  font-family: var(--sans);
+  font-size: calc(var(--ui-font-size) - 1.5px);
+  cursor: pointer;
+}
+.act:hover { background: var(--soft); border-color: var(--bd); }
+.act.signin { background: var(--blue); color: var(--bg); border-color: var(--blue); }
+.act.signin:hover { background: var(--blue2); }
+.switch {
+  position: relative;
+  flex: none;
+  width: 40px;
+  height: 22px;
+  padding: 0;
+  border: 1px solid var(--line);
+  border-radius: 999px;
+  background: var(--panel2);
+  cursor: pointer;
+  transition: background 0.16s;
+}
+.switch.on { background: var(--blue); border-color: var(--blue); }
+.switch:disabled { opacity: 0.5; cursor: not-allowed; }
+.knob {
+  position: absolute;
+  top: 1px;
+  left: 1px;
+  width: 18px;
+  height: 18px;
+  border-radius: 50%;
+  background: var(--bg);
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.2);
+  transition: transform 0.16s;
+}
+.switch.on .knob { transform: translateX(18px); }
+.switch.sm { width: 30px; height: 17px; }
+.switch.sm .knob { width: 13px; height: 13px; }
+.switch.sm.on .knob { transform: translateX(13px); }
+.tag {
+  flex: none;
+  padding: 1px 6px;
+  border-radius: 5px;
+  background: var(--soft);
+  color: var(--muted);
+  font-size: calc(var(--ui-font-size) - 3px);
+}
+.dot {
+  flex: none;
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  background: var(--faint);
+}
+.dot.s-connected { background: var(--ok); }
+.dot.s-connecting { background: var(--warn); }
+.dot.s-error { background: var(--err); }
+.page-title {
+  margin: 0 0 6px;
+  color: var(--ink);
+  font-size: calc(var(--ui-font-size) + 10px);
+  font-weight: 700;
+}
+.page-search {
+  box-sizing: border-box;
+  width: 100%;
+  margin: 0 0 12px;
+  padding: 7px 10px;
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  background: var(--bg);
+  color: var(--ink);
+  font-family: var(--sans);
+  font-size: var(--ui-font-size);
+}
+.listing-count {
+  margin: 0 0 8px;
+  color: var(--faint);
+  font-size: calc(var(--ui-font-size) - 2px);
+}
+.listing { display: flex; flex-direction: column; gap: 2px; }
+.listing-head {
+  margin: 14px 0 6px;
+  color: var(--muted);
+  font-size: calc(var(--ui-font-size) - 2px);
+  font-weight: 600;
+}
+/* Rows are flat rather than carded: the leading glyph, the name and the right
+   action cluster carry the structure, so a border would only add noise. */
+.listing-row {
+  display: flex;
+  flex-direction: column;
+  padding: 5px 0;
+}
+.listing-top {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  min-width: 0;
+}
+/* Only the content fades when an entry is off — the switch that turns it back
+   on has to stay at full contrast. */
+.listing-row.off .listing-main { opacity: 0.5; }
+.listing-main {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  flex: 1;
+  min-width: 0;
+}
+.row-actions {
+  display: flex;
+  flex: none;
+  align-items: center;
+  gap: 4px;
+}
+.listing-glyph {
+  flex: none;
+  width: 15px;
+  height: 15px;
+  color: var(--faint);
+}
+.listing-name {
+  flex: none;
+  font-weight: 500;
+  color: var(--ink);
+}
+.listing-meta {
+  flex: none;
+  font-size: calc(var(--ui-font-size) - 2px);
+  color: var(--faint);
+}
+.listing-desc,
+.listing-path,
+.listing-error {
+  margin: 2px 0 0;
+  font-size: calc(var(--ui-font-size) - 2px);
+  color: var(--muted);
+}
+/* Inside a row the description shares the line with the name, so it truncates
+   instead of wrapping and pushes the meta text to the right edge. */
+.listing-main .listing-desc {
+  margin: 0;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  color: var(--faint);
+}
+.listing-main .listing-meta { margin-left: auto; }
+.listing-path {
+  padding-left: 27px;
+  color: var(--faint);
+  word-break: break-all;
+}
+.listing-error { padding-left: 27px; }
+.listing-indent { padding-left: 27px; }
+
+/* Ghost icon button for the per-row actions — no chrome until hover. */
+.icon-btn {
+  display: flex;
+  flex: none;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  height: 24px;
+  padding: 0;
+  border: none;
+  border-radius: 7px;
+  background: transparent;
+  color: var(--faint);
+  cursor: pointer;
+}
+.icon-btn svg { width: 15px; height: 15px; }
+.icon-btn:hover { background: var(--soft); color: var(--ink); }
+.listing-error {
+  color: var(--err);
+}
+.stat-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+  gap: 8px;
+  margin-bottom: 6px;
+}
+.stat-card {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 12px;
+  border: 1px solid var(--line);
+  border-radius: 10px;
+  background: var(--panel);
+}
+.stat-label { color: var(--muted); font-size: calc(var(--ui-font-size) - 2px); }
+.stat-value { color: var(--ink); font-size: calc(var(--ui-font-size) + 8px); font-weight: 700; }
+.usage-bar {
+  height: 4px;
+  margin-top: 6px;
+  overflow: hidden;
+  border-radius: 999px;
+  background: var(--line2);
+}
+.usage-bar span { display: block; height: 100%; background: var(--blue); }
+.mono { font-family: var(--mono); }
+
+@media (max-width: 640px) {
+  .row { align-items: flex-start; flex-direction: column; }
+}
